ajout des frais d'inscription ainsi que la mise en place d'une classe Semestre valeur passé en long ajout de different indicateur

+++ src/Uttilities/ElementObserver.java
@@ -1,5 +1,5 @@
 package Uttilities;
 
 public interface ElementObserver {
-    public abstract void update(int value);
+    public abstract void update(long value);
 }
